Rust: Miscellaneous Fixes
- Update spec to simplify width
- Add example using the width spec
- Add log to tests
- Show warnings on test failures
- Refactor pipe separated access strings into own module
