+++ rust/osm2lanes/Cargo.toml
@@ -27,9 +27,10 @@ overpass = ["reqwest", "reqwest/blocking", "reqwest/json"]
 tests = ["serde_yaml"]
 
 [dev-dependencies]
-osm2lanes = { path = ".", features = ["tests"] }
 assert-json-diff = "2.0"
 criterion = { version = "0.3", features = ["html_reports"] }
+env_logger = "0.9"
+osm2lanes = { path = ".", features = ["tests"] }
 serde_json = "1"
 
 [lib]
