+++ rust/osm2lanes/src/transform/tags_to_lanes/modes/mod.rs
@@ -1,11 +1,16 @@
 /// Modes of travel
+///
 mod bicycle;
 pub(super) use bicycle::bicycle;
+
 mod bus;
 pub(super) use bus::bus;
+
 mod foot_shoulder;
 pub(super) use foot_shoulder::foot_and_shoulder;
+
 mod parking;
 pub(super) use parking::parking;
+
 mod non_motorized;
 pub(super) use non_motorized::non_motorized;
